LayoutTests/fast/scrolling/ios/scrollto-does-not-interrupt-momentum-scroll-main-frame-expected.txt

Verifies that calling window.scrollTo() during a momentum scroll does not interrupt the momentum scroll in the main frame.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.scrollingElement.scrollTop > 500 became true
PASS document.scrollingElement.scrollTop > scrollTopBeforeScrollTo is true
PASS successfullyParsed is true

TEST COMPLETE
